optimization_service/requirements.txt
fastapi
uvicorn
influxdb-client
python-jose[cryptography]
pandas
numpy
python-multipart
pytest
pytest-cov
